+++ src/dog_marker/database/schemas/entry.py
@@ -1,11 +1,12 @@
 from datetime import datetime
 from uuid import UUID
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
 from dog_marker.dtypes.coordinate import Longitude, Latitude
 
 from .warning_level import WarningLevel
+from .category import Category
 
 
 class Entry(BaseModel):
@@ -18,5 +19,6 @@ class Entry(BaseModel):
     longitude: Longitude
     latitude: Latitude
     warning_level: WarningLevel = WarningLevel.information
+    categories: list[Category] = Field(default_factory=list)
     create_date: datetime
     update_date: datetime